requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
exceptiongroup==1.1.3
fastapi==0.103.2
ffmpeg-python==0.2.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.288
langsmith==0.0.37
llvmlite==0.40.1
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.57.1
numexpr==2.8.6
numpy==1.24.4
openai==0.28.0
openai-whisper==20230314
packaging==23.1
passlib==1.7.4
playsound==1.3.0
PyAudio==0.2.13
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyobjc==9.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
SQLAlchemy==2.0.20
starlette==0.27.0
stripe==6.7.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.3.1
tokenizers==0.14.0
torch==2.0.1
tqdm==4.66.1
twilio==8.8.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
webrtcvad==2.0.10
Werkzeug==2.3.7
whisper==1.1.10
yarl==1.9.2
